Allow using `rand.Source` to initialise
It is currently possible to create a consistent generator by using
`NewWithSeed` which accepts an integer seed value.

Often times we already have a seed source but this package doesn't allow
to make use of it.

So this change adds another way to create a generator that accepts a
more generic seed source.